requirements.txt
amqp==1.4.8
anyjson==0.3.3
billiard==3.3.0.22
celery==3.1.19
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.3
django-allauth==0.23.0
django-celery==3.1.17
django-crispy-forms==1.4.0
django-filter==0.10.0
django-rest-auth==0.4.0
djangorestframework==3.1.3
google-api-python-client==1.4.2
gunicorn==19.3.0
httplib2==0.9.2
kombu==3.0.32
Markdown==2.6.2
numpy==1.10.0.post2
oauth2client==1.5.1
oauthlib==1.0.1
psycopg2==2.6.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
Pygments==2.0.2
pystache==0.5.4
python-memcached==1.57
python-openid==2.2.5
pytz==2015.7
requests==2.7.0
requests-oauthlib==0.5.0
rsa==3.2
scikit-learn==0.16.1
scipy==0.16.0
simplejson==3.8.0
six==1.9.0
sklearn==0.0
static==1.1.1
static3==0.6.1
uritemplate==0.6
wheel==0.24.0
